more LoginPage refactor + loginPage.interfaces.ts + useLoginMutation
