+++ src/pages/login/loginPage.interfaces.ts
@@ -1,0 +1,4 @@
+export interface Inputs {
+  email: string
+  password: string
+}
